refactor: add color initials to avatar, improve usage of user avatar
