+++ apps/nextjs/src/app/[locale]/manage/users/create/_components/stepper-navigation.tsx
