+++ apps/nextjs/src/components/layout/header/user.tsx
@@ -1,13 +1,13 @@
 import { UnstyledButton } from "@mantine/core";
 
-import { UserAvatar } from "~/components/user-avatar";
+import { CurrentUserAvatar } from "~/components/user-avatar";
 import { UserAvatarMenu } from "~/components/user-avatar-menu";
 
 export const UserButton = () => {
   return (
     <UserAvatarMenu>
       <UnstyledButton>
-        <UserAvatar size="md" />
+        <CurrentUserAvatar size="md" />
       </UnstyledButton>
     </UserAvatarMenu>
   );
